+++ app/src/main/java/com/eagskunst/emmanuel/gamingnews/views/SettingsActivity.kt
@@ -5,6 +5,9 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.runtime.getValue
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.eagskunst.emmanuel.gamingnews.ui.settings.SettingsScreen
 import com.eagskunst.emmanuel.gamingnews.ui.settings.SettingsViewModel
 import com.eagskunst.emmanuel.gamingnews.ui.theme.GamingNewsTheme
@@ -19,7 +22,8 @@ class SettingsActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            GamingNewsTheme {
+            val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+            GamingNewsTheme(darkTheme = uiState.darkTheme) {
                 SettingsScreen(
                     viewModel = viewModel,
                     onBackClick = { finish() }